requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
requests==2.31.0
python-dotenv==1.0.0
pytest==8.2.0
pytest-mock==3.14.0
tenacity==8.3.0
pyarrow>=16.0.0
pandas>=2.2.2
